+++ messages/de/structure.php
@@ -137,6 +137,8 @@ return [
     'STATUSV_ACCEPTED'                    => 'Annehmen',
     'STATUS_REJECTED'                     => 'Abgelehnt',
     'STATUSV_REJECTED'                    => 'Ablehnen',
+    'STATUS_QUORUM_MISSED'                => 'Quorum verfehlt',
+    'STATUS_QUORUM_REACHED'               => 'Quorum erreicht',
     'STATUS_MODIFIED_ACCEPTED'            => 'Modifiziert übernommen',
     'STATUSV_MODIFIED_ACCEPTED'           => 'Modifiziert übernehmen',
     'STATUS_MODIFIED'                     => 'Modifiziert',