+++ views/voting/voting-block.vue.php
@@ -1,8 +1,8 @@
 <?php
 
-
 use app\components\UrlHelper;
 use app\models\layoutHooks\Layout;
+use app\models\quorumType\IQuorumType;
 use app\models\votings\AnswerTemplates;
 use app\models\db\{Consultation, ConsultationUserGroup, IMotion, User, VotingBlock};
 use yii\helpers\Html;
@@ -108,7 +108,7 @@ ob_start();
                     }
                     ?>
                 </div>
-                <div class="result" v-if="isClosed && votingHasMajority">
+                <div class="result" v-if="isClosed && (votingHasMajority || votingHasQuorum)">
                     <div class="accepted" v-if="itemIsAccepted(groupedVoting)">
                         <span class="glyphicon glyphicon-ok" aria-hidden="true"></span>
                         <?= Yii::t('voting', 'status_accepted') ?>
@@ -116,6 +116,14 @@ ob_start();
                     <div class="rejected" v-if="itemIsRejected(groupedVoting)">
                         <span class="glyphicon glyphicon-minus" aria-hidden="true"></span>
                         <?= Yii::t('voting', 'status_rejected') ?>
+                    </div>
+                    <div class="accepted" v-if="itemIsQuorumReached(groupedVoting)">
+                        <span class="glyphicon glyphicon-ok" aria-hidden="true"></span>
+                        <?= Yii::t('voting', 'status_quorum_reached') ?>
+                    </div>
+                    <div class="rejected" v-if="itemIsQuorumFailed(groupedVoting)">
+                        <span class="glyphicon glyphicon-minus" aria-hidden="true"></span>
+                        <?= Yii::t('voting', 'status_quorum_missed') ?>
                     </div>
                 </div>
             </li>
@@ -160,21 +168,6 @@ $html = ob_get_clean();
     const RESULTS_PUBLIC_YES = <?= VotingBlock::RESULTS_PUBLIC_YES ?>;
     const RESULTS_PUBLIC_NO = <?= VotingBlock::RESULTS_PUBLIC_NO ?>;
 
-    // The voting is not performed using Antragsgrün
-    const STATUS_OFFLINE = <?= VotingBlock::STATUS_OFFLINE ?>;
-
-    // Votings that have been created and will be using Antragsgrün, but are not active yet
-    const STATUS_PREPARING = <?= VotingBlock::STATUS_PREPARING ?>;
-
-    // Currently open for voting.
-    const STATUS_OPEN = <?= VotingBlock::STATUS_OPEN ?>;
-
-    // Vorting is closed.
-    const STATUS_CLOSED = <?= VotingBlock::STATUS_CLOSED ?>;
-
-    const VOTING_STATUS_ACCEPTED = <?= IMotion::STATUS_ACCEPTED ?>;
-    const VOTING_STATUS_REJECTED = <?= IMotion::STATUS_REJECTED ?>;
-
     const ANSWER_TEMPLATE_YES_NO_ABSTENTION = <?= AnswerTemplates::TEMPLATE_YES_NO_ABSTENTION ?>;
     const ANSWER_TEMPLATE_YES_NO = <?= AnswerTemplates::TEMPLATE_YES_NO ?>;
     const ANSWER_TEMPLATE_PRESENT = <?= AnswerTemplates::TEMPLATE_PRESENT ?>;
@@ -189,10 +182,6 @@ $html = ob_get_clean();
             }
         },
         computed: {
-            votingHasMajority: function () {
-                // Used for the currently running vote as it is
-                return this.voting.answers_template === ANSWER_TEMPLATE_YES_NO_ABSTENTION || this.answers_template === ANSWER_TEMPLATE_YES_NO;
-            },
             votingOptionButtons: function () {
                 return this.voting.answers.map((answer) => {
                     return this.voteAnswerToCss(answer);
@@ -210,12 +199,6 @@ $html = ob_get_clean();
             resultsPublic: function () {
                 return this.voting.results_public === RESULTS_PUBLIC_YES;
             },
-            isOpen: function () {
-                return this.voting.status === STATUS_OPEN;
-            },
-            isClosed: function () {
-                return this.voting.status === STATUS_CLOSED;
-            }
         },
         methods: {
             vote: function (groupedVoting, voteOption) {
@@ -250,27 +233,6 @@ $html = ob_get_clean();
             votedOption: function (group) {
                 const answer = this.getVoteOptionById(group.voted);
                 return this.voteAnswerToCss(answer);
-            },
-            itemIsAccepted: function (groupedItem) {
-                return groupedItem[0].voting_status === VOTING_STATUS_ACCEPTED;
-            },
-            itemIsRejected: function (groupedItem) {
-                return groupedItem[0].voting_status === VOTING_STATUS_REJECTED;
-            },
-            hasVoteList: function (groupedItem) {
-                return groupedItem[0].votes !== undefined;
-            },
-            isVoteListShown: function (groupedItem) {
-                const showId = groupedItem[0].type + '-' + groupedItem[0].id;
-                return this.shownVoteLists.indexOf(showId) !== -1;
-            },
-            showVoteList: function (groupedItem) {
-                const showId = groupedItem[0].type + '-' + groupedItem[0].id;
-                this.shownVoteLists.push(showId);
-            },
-            hideVoteList: function (groupedItem) {
-                const hideId = groupedItem[0].type + '-' + groupedItem[0].id;
-                this.shownVoteLists = this.shownVoteLists.filter(id => id !== hideId);
             }
         }
     });